+++ spec/unit/zmanim/util/geo_location_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+require 'time'
 
 describe Zmanim::Util::GeoLocation do
   describe '.GMT' do